wip call scanblocks from add not ls
